client/requirements.txt
pygame==2.5.2
python-socketio[client]==5.11.0
websocket-client==1.7.0
requests==2.31.0
schedule==1.2.1
pystray==0.19.5
Pillow==10.2.0
